+++ Meals_Schdueler2/app/src/main/java/com/example/meals_schdueler/EditWeeklyDialog.kt
@@ -683,7 +683,7 @@ class EditWeeklyDialog(
 
             var ingredientList: ArrayList<Ingredient> = ArrayList()
 
-            var quantities: HashMap<Int, ArrayList<Int>> = HashMap()
+            var quantities: HashMap<Int, ArrayList<Float>> = HashMap()
 
             var ids: HashMap<Int, ArrayList<Int>> = HashMap()
 
@@ -718,18 +718,18 @@ class EditWeeklyDialog(
 
                     if (!recipeIngredientMap.containsKey(currentIngId)) {
                         var recipeIngredients: ArrayList<Ingredient> = ArrayList()
-                        var quantitiy: ArrayList<Int> = ArrayList()
+                        var quantitiy: ArrayList<Float> = ArrayList()
                         var idss: ArrayList<Int> = ArrayList()
                         recipeIngredientMap.put(currentIngId, recipeIngredients)
                         recipeIngredientMap.get(currentIngId)!!.add(ing)
                         quantities.put(currentIngId, quantitiy)
-                        quantities.get(currentIngId)!!.add(recipesAndIngredients2[14].toInt())
+                        quantities.get(currentIngId)!!.add(recipesAndIngredients2[14].toFloat())
                         ids.put(currentIngId, idss)
                         ids.get(currentIngId)!!.add(recipesAndIngredients2[0].toInt())
 
                     } else {
                         recipeIngredientMap.get(currentIngId)!!.add(ing)
-                        quantities.get(currentIngId)!!.add(recipesAndIngredients2[14].toInt())
+                        quantities.get(currentIngId)!!.add(recipesAndIngredients2[14].toFloat())
                         ids.get(currentIngId)!!.add(recipesAndIngredients2[0].toInt())
                     }
