+++ Meals_Schdueler2/app/src/main/java/com/example/meals_schdueler/LoadingData.kt
@@ -168,13 +168,13 @@ class LoadingData : AppCompatActivity(), GetAndPost {
             // first we extract the ids and quantity into map.
             // then we use another map to convert all ids to real ingredients.
 
-            var hashMap: HashMap<Int, Pair<ArrayList<String>, ArrayList<Int>>> =
+            var hashMap: HashMap<Int, Pair<ArrayList<String>, ArrayList<Float>>> =
                 HashMap()
 
             var map: HashMap<Int, ArrayList<Ingredient>> = HashMap()
 
 
-            var quantities: ArrayList<Int> = ArrayList()
+            var quantities: ArrayList<Float> = ArrayList()
             var ids: ArrayList<String> = ArrayList()
             var ingredientList2: ArrayList<Ingredient> = ArrayList()
             //  var quantities: ArrayList<String> = ArrayList()
@@ -209,7 +209,7 @@ class LoadingData : AppCompatActivity(), GetAndPost {
                     }
                     if (ingID != recipesAndIngredients2[9]) {
                         ids.add(recipesAndIngredients2[9])
-                        quantities.add(recipesAndIngredients2[10].toInt())
+                        quantities.add(recipesAndIngredients2[10].toFloat())
                         // getting the specific ingredient by its ID
                         ingredientList2.add(
                             UserPropertiesSingelton.getInstance()!!.getUserIngredientss()
@@ -251,7 +251,7 @@ class LoadingData : AppCompatActivity(), GetAndPost {
                                 recipe2[7].toBoolean(),
                                 recipe2[8].toBoolean(),
                                 map.get(recipe2[0].toInt())!!,
-                                hashMap.get(recipe2[0].toInt())!!.second
+                                hashMap.get(recipe2[0].toFloat())!!.second
 
                             )
                         )
