+++ Meals_Schdueler2/app/src/main/java/com/example/meals_schdueler/MyRecipeIngredietns.kt
@@ -18,7 +18,7 @@ class MyRecipeIngredietns(
     recipeName: String,
     pictureBitMap: Bitmap,
     numOfPortions: String,
-    quantityList: ArrayList<Int>,
+    quantityList: ArrayList<Float>,
     totalCost: Double
 ) : DialogFragment() {
