+++ Meals_Schdueler2/app/src/main/java/com/example/meals_schdueler/My_Ingredients_Recipe_RecyclerViewAdapter.kt
@@ -18,7 +18,7 @@ import kotlin.collections.ArrayList
  */
 class My_Ingredients_Recipe_RecyclerViewAdapter(
     private var values: ArrayList<Ingredient>,
-    quantityList: ArrayList<Int>,
+    quantityList: ArrayList<Float>,
     childFragmentManager: FragmentManager
 ) : RecyclerView.Adapter<My_Ingredients_Recipe_RecyclerViewAdapter.ViewHolder>() {
